requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==4.2.0
attrs==23.2.0
blinker==1.7.0
bs4
cachetools==5.3.2
<<<<<<< HEAD
certifi==2023.11.17
charset-normalizer==3.3.2
chardet
=======
chardet
certifi==2023.11.17
charset-normalizer==3.3.2
>>>>>>> main
click==8.1.7
dataclasses-json==0.6.3
distro==1.9.0
frozenlist==1.4.1
faiss-cpu
gitdb==4.0.11
gtts
GitPython==3.1.41
google-ai-generativelanguage==0.4.0
google-api-core==2.15.0
google-auth==2.26.2
google-generativeai==0.3.2
googleapis-common-protos==1.62.0
greenlet==3.0.3
grpcio==1.60.0
grpcio-status==1.60.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
importlib-metadata==7.0.1
Jinja2==3.1.3
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.0
jsonschema-specifications==2023.12.1
langchain==0.1.1
langchain-community==0.0.13
langchain_google_genai
langchain-core==0.1.12
langsmith==0.0.83
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.2
mdurl==0.1.2
multidict==6.0.4
<<<<<<< HEAD
matplotlib
mypy-extensions==1.0.0
numpy==1.26.3
openpyxl
packaging==23.2
pandas==2.1.4
pefile
Pillow==10.2.0
plotly==5.17.0
=======
mypy-extensions==1.0.0
numpy==1.26.3
openai==1.8.0
packaging==23.2
pandas==2.1.4
Pillow==10.2.0
>>>>>>> main
proto-plus==1.23.0
protobuf==4.25.2
pyarrow==14.0.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.5.3
pydantic_core==2.14.6
pydeck==0.8.1b0
PyPDF2
Pygments==2.17.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.1
requests==2.31.0
rich==13.7.0
rpds-py==0.17.1
rsa==4.9
six==1.16.0
smmap==5.0.1
<<<<<<< HEAD
seaborn
streamlit_authenticator
sniffio==1.3.0
SQLAlchemy==2.0.25
streamlit
=======
sniffio==1.3.0
SQLAlchemy==2.0.25
streamlit==1.34.0
>>>>>>> main
tenacity==8.2.3
toml==0.10.2
toolz==0.12.0
tornado==6.4
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
urllib3==2.1.0
validators==0.22.0
watchdog==3.0.0
yarl==1.9.4
zipp==3.17.0
SpeechRecognition
pydub
pyaudio
